+++ CHANGELOG.md
@@ -13,6 +13,9 @@ This project does not follow SemVer, since modules are independent of each other
 ### slack/sentry
 - Add Slack notifier for Sentry issue alerts [#184](https://github.com/dbl-works/terraform/pull/184)
 
+### stack/app
+- Add Add elasticache_name as variables [#187](https://github.com/dbl-works/terraform/pull/187)
+
 ## [v2023.03.06]
 ### iam/iam-policy-for-secrets
 - Added a new module to manage access to the secretsmaanger and kms based on project_access variables. [#175](https://github.com/dbl-works/terraform/pull/175)